move TPlanPrinter to header
