+++ ydb/library/yql/providers/dq/provider/yql_dq_gateway.cpp
@@ -1,6 +1,7 @@
 #include "yql_dq_gateway.h"
 
 #include <yql/essentials/providers/common/provider/yql_provider_names.h>
+#include <ydb/library/yql/dq/tasks/dq_tasks_printer.h>
 #include <ydb/library/yql/providers/dq/api/grpc/api.grpc.pb.h>
 #include <ydb/library/yql/providers/dq/common/yql_dq_common.h>
 #include <ydb/library/yql/providers/dq/actors/proto_builder.h>
@@ -26,96 +27,6 @@
 namespace NYql {
 
 using namespace NThreading;
-
-class TPlanPrinter {
-public:
-    TStringBuilder b;
-
-    void DescribeChannel(const auto& ch, bool spilling) {
-        if (spilling) {
-            b << "Ch" << ch.GetId() << " [shape=diamond, label=\"Ch" << ch.GetId() << "\", color=\"red\"];";
-        } else {
-            b << "Ch" << ch.GetId() << " [shape=diamond, label=\"Ch" << ch.GetId() << "\"];";
-        }
-    }
-
-    void PrintInputChannel(const auto& ch, const auto& type) {
-        b << "Ch" << ch.GetId() << " -> T" << ch.GetDstTaskId() << " [label=" << "\"" << type << "\"];\n";
-    }
-
-    void PrintOutputChannel(const auto& ch, const auto& type) {
-        b << "T" << ch.GetSrcTaskId() << " -> Ch" << ch.GetId() << " [label=" << "\"" << type << "\"];\n";
-    }
-
-    void PrintSource(auto taskId, auto sourceIndex) {
-        b << "S" << taskId << "_" << sourceIndex << " -> T" << taskId << " [label=" << "\"S" << sourceIndex << "\"];\n";
-    }
-
-    void DescribeSource(auto taskId, auto sourceIndex) {
-        b << "S" << taskId << "_" << sourceIndex << " ";
-        b << "[shape=square, label=\"" << taskId << "/" << sourceIndex << "\"];\n";
-    }
-
-    void PrintTask(const auto& task) {
-        int index = 0;
-        for (const auto& input : task.GetInputs()) {
-            TString inputName = "Unknown";
-            bool isSource = false;
-            if (input.HasUnionAll()) { inputName = "UnionAll"; }
-            else if (input.HasMerge()) { inputName = "Merge"; }
-            else if (input.HasSource()) { inputName = "Source"; isSource = true; }
-            if (isSource) {
-                PrintSource(task.GetId(), index);
-            } else {
-                for (const auto& ch : input.GetChannels()) {
-                    PrintInputChannel(ch, inputName);
-                }
-            }
-            index ++;
-        }
-        for (const auto& output : task.GetOutputs()) {
-            TString outputName = "Unknown";
-            if (output.HasMap()) { outputName = "Map"; }
-            else if (output.HasRangePartition()) { outputName = "Range"; }
-            else if (output.HasHashPartition()) { outputName = "Hash"; }
-            else if (output.HasBroadcast()) { outputName = "Broadcast"; }
-            // TODO: effects, sink
-            for (const auto& ch : output.GetChannels()) {
-                PrintOutputChannel(ch, outputName);
-            }
-        }
-    }
-
-    void DescribeTask(const auto& task) {
-        b << "T" << task.GetId() << " [shape=circle, label=\"" << task.GetId() << "/" << task.GetStageId() << "\"];\n";
-        int index = 0;
-        for (const auto& input : task.GetInputs()) {
-            if (input.HasSource()) {
-                DescribeSource(task.GetId(), index);
-            }
-            index ++;
-        }
-        for (const auto& output : task.GetOutputs()) {
-            for (const auto& ch : output.GetChannels()) {
-                DescribeChannel(ch, task.GetEnableSpilling());
-            }
-        }
-    }
- 
-    TString Print(const NDqs::TPlan& plan) {
-        b.clear();
-        b << "digraph G {\n";
-        for (const auto& task : plan.Tasks) {
-            DescribeTask(task);
-        }
-        b << "\n";
-        for (const auto& task : plan.Tasks) {
-            PrintTask(task);
-        }
-        b << "}\n";
-        return b;
-    }
-};
 
 class TDqTaskScheduler : public TTaskScheduler {
 private:
@@ -390,7 +301,7 @@ public:
         resultFormatSettings.SizeLimit = settings->_AllResultsBytesLimit.Get();
         resultFormatSettings.RowsLimit = settings->_RowsLimitPerWrite.Get();
 
-        YQL_CLOG(TRACE, ProviderDq) << TPlanPrinter().Print(plan);
+        YQL_CLOG(TRACE, ProviderDq) << TPlanPrinter().Print(plan.Tasks);
 
         {
             auto& secParams = *queryPB.MutableSecureParams();
